Add OnSubscriptionNotificationCore trait
This was requested at some point and makes some sense. Add a way to get
callbacks directly when a NotificationMessage arrives on the client,
instead of once per monitored item. This might be useful for batching.

This change should actually be mostly backwards compatible. I blanket
implemented the new trait using the old OnSubscriptionNotification
trait.
