+++ async-opcua-client/src/lib.rs
@@ -125,9 +125,10 @@ pub use config::{ClientConfig, ClientEndpoint, ClientUserToken, ANONYMOUS_USER_T
 pub use retry::{ExponentialBackoff, SessionRetryPolicy};
 pub use session::{
     Client, DataChangeCallback, DefaultRetryPolicy, EventCallback, HistoryReadAction,
-    HistoryUpdateAction, MonitoredItem, OnSubscriptionNotification, RequestRetryPolicy, Session,
-    SessionActivity, SessionBuilder, SessionConnectMode, SessionEventLoop, SessionPollResult,
-    Subscription, SubscriptionActivity, SubscriptionCallbacks, UARequest,
+    HistoryUpdateAction, MonitoredItem, OnSubscriptionNotification, OnSubscriptionNotificationCore,
+    RequestRetryPolicy, Session, SessionActivity, SessionBuilder, SessionConnectMode,
+    SessionEventLoop, SessionPollResult, Subscription, SubscriptionActivity, SubscriptionCallbacks,
+    UARequest,
 };
 pub use transport::AsyncSecureChannel;
